+++ vidrioPropiedades.py
@@ -6,24 +6,24 @@ class VidrioPropiedades:
 
 
     #propiedades del vidrio
-    T = Decimal(0.84) #transmitancia
-    absortancia = Decimal(0) # absortancia
-    ev = Decimal(0.8) #Emisividad
-    alphav = Decimal(0.006)
+    T = float(0.84) #transmitancia
+    absortancia = float(0) # absortancia
+    ev = float(0.8) #Emisividad
+    alphav = float(0.006)
 
 
     #propiedades vidrio - aire
 
-    ufva = Decimal(0) #viscocidad cinematica del vidrio-aire
-    densva = Decimal(0) # densidad del vidrio aire
-    condva = Decimal(0) # conductividad del vidrio aire
-    cpva = Decimal(0)  # calor especifico
-    betava = Decimal(0)
-    deltava = Decimal(0)
-    prva = Decimal(0)
-    vfva  = Decimal(0)
-    grva = Decimal(0)
-    rava = Decimal(0)
-    nusseltva  = Decimal(0)
-    hg = Decimal(0)
-    tmva = Decimal(0) #temp media
+    ufva = float(0) #viscocidad cinematica del vidrio-aire
+    densva = float(0) # densidad del vidrio aire
+    condva = float(0) # conductividad del vidrio aire
+    cpva = float(0)  # calor especifico
+    betava = float(0)
+    deltava = float(0)
+    prva = float(0)
+    vfva  = float(0)
+    grva = float(0)
+    rava = float(0)
+    nusseltva  = float(0)
+    hg = float(0)
+    tmva = float(0) #temp media
